chore: Consolidate TurboJson tests into turborepo-turbo-json crate
Move duplicate and movable tests from turborepo-lib/turbo_json/mod.rs
to turborepo-turbo-json/lib.rs to continue modularization.

- Remove 14 duplicate tests that already existed in turborepo-turbo-json
- Move 11 unique tests to turborepo-turbo-json (boundaries, task outputs,
  output logs mode, daemon, UI serialization, config validation, etc.)
- Keep only test_deserialize_task_definition in turborepo-lib (requires
  TaskDefinition::from_raw from turborepo-engine, which cannot be a
  dependency of turborepo-turbo-json due to reverse dependency)
- Copy snapshot files with updated naming convention

This reduces turborepo-lib/turbo_json/mod.rs from ~825 lines to ~280 lines.
