+++ crates/turborepo-turbo-json/Cargo.toml
@@ -31,6 +31,7 @@ turborepo-types = { workspace = true }
 turborepo-unescape = { workspace = true }
 
 [dev-dependencies]
+anyhow = { workspace = true }
 insta = { workspace = true }
 pretty_assertions = { workspace = true }
 tempfile = { workspace = true }
